add support for .js config files
It's needed for those rare cases when one needs to specify function or
other non-JSON type in settings.

We now use `require` instead of `JSON.parse`. It supports both JSON and
JS files. However `require`, unlike `fs.readFile`, does not resolve
relative paths in the context of current working directory.  We need to
that explicitly.