feat(completion): allow to get completions for any string, not just `process.argv`
Add new method `.getCompletion(args, done)` which receives an array of strings
representing a line to complete, and calls the `done` callback with the
resulting completions.

Update the completion tests to pass the arguments explicitly to yargs. We can't
pass them by mocking `process.argv` now because `process.argv` is only checked
on `index.js`, which is called before the tests can apply the mock.