fix: for args that have skipValidation set to `true`, check if the parsed arg is `true` (#619)
* For args that have skipValidation set to `true`, check if the parsed arg is `true`

When an arg is defined, it will always exist in the parsed argv. Thus, validation would always be skipped, regardless of whether an arg that has `skipValidation` set to `true` is actually passed on the command line. This commit checks to see if the parsed argv value is actually set to true.

* Added test for `skipValidation` fix