fix: $0 contains first arg on bundled electron app
Default behavior of yargs is to take the first two args of process.argv
to put into $0 as this is usually the program to execute a script
e.g. 'node app.js' but for bundled electron apps this results in $0
being populated with not only the executing program but also the first
parameter. This fix modifies the check for node/iojs (and adds electron)
so in case of those 3, the executing program is skipped in $0 as it was
before, but in case where a bundled app is executed, the first parameter
is skipped.

This is only a cosmetic change for proper help printouts which use $0
A manual yargs.parse(process.argv.slice(1)) is still necessary for
electron not to skip the first parameter, during parsing