Removed "default: false" on verbose.
Removed "default: false" on verbose. I was mildly confused when I copied the example and called it with `-v`, expecting no value to default the value to true. However it appears as if the default value overrides both the "not passed" value, and the "passed with no value" parameter. This makes me believe "default: false" should never be used, as the default of default is effectively false.

At the very least I believe it shouldn't be part of the example in the README, especially on conjunction with the verbose parameter, which makes a program that uses verbose differently than any program I have ever seen.