Some clarifying updates to de translation
Although I'm not a *native* German speaker (I *am* a native English
    speaker though), there were two specific
things that stood out in the translation as strange. (I will double
    check with native speakers if I have the chance -- and any native
    speakers, please share your thoughts.)

I did do some double and triple checking on word usage and other
translations, as well.

@maxrimue - I'm guessing that you're a native speaker of German, and you
contributed to the translation in the past - perhaps if you have a
minute you can review this?

* "count" - Based on the way that this would be used, I'm looking at it
as "a count of something" ... "a quantity of something" ... and for that
I think "Anzahl" would be a better choice. "Zähler" would be more of "a
counter" or "something that counts" (or even "someone who counts").

* "choices" - "Möglichkeiten" would more literally translate as
"possibilities" (to say "es ist möglich" is to say "it is
    possible/capable of happening/occurring"). But that's kinda what it
literally means - possibilities. "Auswahl" (and the verb auswählen - to
    choose, to select) would be a choice from a set of options. And
"optionen" would literally be the plural of "options"... I *suppose* one
could write "Auswahlmöglichkeiten" ... but to my ear and understanding,
      that would refer to the **availability** of **an opportunity to
      select something** not the **available selection options**
      themselves.