fix(strictCommands): report unknown argument instead of unknown command if no subcommands configured
This gives user a friendlier error message when unexpected positional argument is passed to the command. It does not really make sense to say "Unknown command: foo" - there are not subcommands to this command. This is an unknown argument and should be reported as such.

A more correct solution would probably be to not report an error at all for this case under strictCommands, but report it under strict mode instead. But this will be a breaking change as it is explicitly [stated in the docs](https://yargs.js.org/docs/#api-strictcommandsenabledtrue) that unknown positional arguments are reported as errors under strictCommands, so current solution should do for now.