Fix Zsh autoloaded completion to not require double tab
Currently when you use the generated completion script in Zsh via autoloading, it doesn't provide completions on the first tab press. This is because when Zsh autoloads a completion script, it expects it to directly generate completions, this little trick allows it to be directly evaluated and used as an autoloaded script.

P.S. I think if the shell function is named the same as the autoloaded script, Zsh will call it directly on subsequent calls, bypassing the rest of the script, e.g. the `if` code below. See https://zsh.sourceforge.io/Doc/Release/Functions.html#Autoloading-Functions

See https://github.com/kislyuk/argcomplete/pull/475